Add user login and registration

+++ studyroom/base/views.py
@@ -1,9 +1,60 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
-from .models import Room, Topic
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, User
 from .forms import RoomForm
 
 # Create your views here.
+
+
+def login_page(request):
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect("home")
+
+    if request.method == "POST":
+        username = request.POST.get("username").lower()
+        password = request.POST.get("password")
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User doesn't exist!")
+
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "Invalid Username or Password!")
+
+    context = {"page": page}
+    return render(request, "base/login_register.html", context)
+
+
+def logout_user(request):
+    logout(request)
+    return redirect("login")
+
+
+def register_user(request):
+    form = UserCreationForm()
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "An error occured during registration")
+
+    context = {"form": form}
+    return render(request, "base/login_register.html", context)
 
 
 def home(request):
@@ -21,6 +72,7 @@ def room(request, pk):
     return render(request, "base/room.html", context)
 
 
+@login_required(login_url="login")
 def create_room(request):
     form = RoomForm()
     if request.method == "POST":
@@ -32,9 +84,15 @@ def create_room(request):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        messages.error(request, "You can not make changes to this room!")
+        return redirect("home")
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -44,8 +102,12 @@ def update_room(request, pk):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        messages.error(request, "You can not make changes to this room!")
+        return redirect("home")
     if request.method == "POST":
         room.delete()
         return redirect("home")
